Data_Collection_Project/requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.32.3
